refactor: repackage post model
